Split API responsabilities.
