--- Reddit/Http/Model/Child.swift
@@ -1,12 +1,0 @@
-//
-//  Child.swift
-//  Reddit
-//
-//  Created by Vanesa Korbenfeld on 10/04/2023.
-//
-
-import Foundation
-
-struct Child: Codable {
-    let data: ChildData
-}
