--- Reddit/Http/Model/Data.swift
@@ -1,12 +1,0 @@
-//
-//  Data.swift
-//  Reddit
-//
-//  Created by Vanesa Korbenfeld on 10/04/2023.
-//
-
-import Foundation
-
-struct Data: Codable {
-    let children: [Child]
-}
