--- Reddit/Http/Model/News.swift
@@ -1,12 +1,0 @@
-//
-//  Reddit.swift
-//  Reddit
-//
-//  Created by Vanesa Korbenfeld on 07/04/2023.
-//
-
-import Foundation
-
-struct News: Codable {
-    let data: Data
-}
