+++ Reddit/ViewController/Error/ErrorManager.swift
